requirements.txt
azure-functions
sendgrid